backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
httpx==0.28.0
pydantic-settings==2.7.0
alembic==1.14.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
